Fix LICENSE.md file to CC-BY-SA, as README.md says
README.md says this repository is under CC-BY-SA, but the LICENSE file
had the text of CC-BY in it.  I'm assuming that README.md is right and
that the LICENSE file content was just an accidental mis-copy
somewhere along the way.  This change switches the license file to
have CC-BY-SA and changes its format to Markdown (and hence the name
from 'LICENSE' to 'LICENSE.md').  The format change is both to match
README.md and to match what seems to be the trend in freely-licensed
projects of using Markdown format for license files.