Add "DPG" acronym at first chance in two modules
This addresses a comment first made (in PR #13) about the Introduction,
but addresses it in the Policy module as well as in the Introduction;
there are currently no other modules where this change is needed.

Note that the expansion is capitalized differently in each place.  I
think that's okay, but even if it were something we wanted to fix,
fixing it would be separate from adding the acronym.