Update slides based on feedback
Un-rearrange so that the slide deck now follows the same outline/TOC
as the toolkit document itself, list key recommendations and expected
outcomes, and use DPGA's slide templates.  Also: have a placeholder
column for examples, but don't include any examples yet as interviews
are still in progress.  These changes are based on this email:

  From: Ibrahim Emad
  Subject: Re: Slide-deck overview of DPG Toolkit.
  To: Karl Fogel
  CC: Sunita Grote,, Victor Grau Serrat, Lucy Harris,
      Jameson Voisin, James Vasile, Amye Scavarda Perrin
  Date: Thu, 12 Aug 2021 14:49:16 +0000
  Message-ID: <AS8PR05MB86605390FD30463B1A3F04F3D8F99\
              @AS8PR05MB8660.eurprd05.prod.outlook.com>