Various independent changes (see below)
This was actually various in-progress changes that Susy had in her
local working tree.  They got accidentally committed when she and Karl
were discussing the finer points of branches vs PRs vs commits and
Susy was trying to show Karl a certain behavior of GitHub Desktop.
The commit first went in with the summary "test", so Karl amended the
commit message and force-pushed it to the 'susyedits' branch.