Massive reformatting, plus some light editing
Most of this change is source reformatting -- specifically:

  * Adding line breaks to long paragraphs so that future edits are
    reviewable as line-based diffs;

  * Reverting some non-working or not-reliably-working links to TBD
    notes (to fix in the upcoming grand link-fixing pass later);

  * Fixing up the formatting of some lists;

  * Adding two spaces after sentence-ending periods (so auto-wrapping
    will work, and so that sentence-wise motion commands will work in
   editing software that supports that);

  * Various other miscellaneous source formatting fixes.

However, I also made a few minor substantive changes along the way.  I
wouldn't normally mix formatting and real changes like that, but the
situation on this branch is special (see PR #39) and compels me to
temporarily forego the usual niceties until the branch is merged.