Start testing intra-document linking methods
This commit (and subsequent commits on the new 'link-tests' branch)
are for testing Markdown linking methods.  We need to use a branch
because we must push to GitHub in order to see whether links work in
GitHub-rendered HTML as well as in the PDF.  Expect more commits :-).

The first paragraph of the main text of "Community And Ecosystem"
module (in community.md) is where I'm testing, because that paragraph
happens to have two intra-document links to two different modules.