Move most periods and all commas outside quotes
Move most periods and all commas that were before a closing
double-quote to after it, except for periods that conclude a sentence
or sentence fragment.  In the latter case, leave the period so as to
end what's inside the quote while avoiding the '.".'  situation.

I won't fill this commit message will all the reasons for this change,
but am happy to discuss it with anyone who's curious.