Remove and ignore some tmp files
These tmp files were created on Mac OS and just accidentally checked
in to Git.  Remove them, and update .gitignore so that they are less
likely to end up in version control in the future (actually one of
them had already been listed in .gitignore since commit f176065b75b,
but it still needed to be removed from the tree).