Handle the "open\nsource" -> "open-source" case
I believe this completes the fix for issue #25.  It finishes what was
started in commit 6901fc077520e and commit c7109a202017.  I forgot to
mention issue #25 in those commits, but I hope that having the issue
number in this log message, and then mentioning those commits here,
will leave enough of an audit trail for future archaeologists.