Improve PDF output via source formatting fixes
This commit consists almost entirely of source formatting fixes --
mainly to list item positioning and vertical spacing -- to make the
PDF render as intended.

In the few places where there is a sort of substantive change, it's
still fundamentally a formatting fix.  For example, we adjust the
wording of the sentence beginning "The figure below..." on p. 67 to
account for the fact that the ecosystem map figure may end up by
itself on p. 68, and we adjust the wording of a URL's link text in the
"Lutece" entry in the examples appendix in the course of correcting
an overally formatting problem with that URL.